refactor(lsp): extract remaining handlers into language modules
Extract remaining LSP handlers from server_impl/mod.rs into focused
language feature modules:

- code_actions.rs: handle_code_action, handle_code_action_resolve,
  handle_code_actions_pragmas
- rename.rs: handle_prepare_rename, handle_rename, handle_rename_workspace,
  get_token_at_position, get_token_bounds, is_valid_identifier
- hierarchy.rs: handle_prepare_type_hierarchy, handle_type_hierarchy_supertypes,
  handle_type_hierarchy_subtypes, handle_prepare_call_hierarchy,
  handle_incoming_calls, handle_outgoing_calls, json_to_call_hierarchy_item
- semantic_tokens.rs: handle_semantic_tokens, handle_semantic_tokens_full,
  handle_semantic_tokens_range
- misc.rs: handle_inlay_hints, handle_document_links, handle_selection_range,
  handle_code_lens, handle_code_lens_resolve, handle_inline_completion,
  handle_inline_value, handle_moniker, handle_document_color,
  handle_color_presentation, handle_linked_editing_range,
  handle_test_discovery, handle_execute_command, count_references_text_based,
  workspace_roots

Dispatch routing unchanged - all handlers still called via self.handle_*().
mod.rs reduced from ~4000 to ~1800 lines.

Part of the ongoing LSP server modularization effort.
