+++ crates/perl-parser/src/lsp/server_impl/language/mod.rs
@@ -7,10 +7,20 @@
 //! - references: Find references and document highlights
 //! - symbols: Document symbols and folding ranges
 //! - formatting: Document and range formatting
+//! - code_actions: Code actions and quick fixes
+//! - rename: Symbol renaming (single file and workspace)
+//! - hierarchy: Type hierarchy and call hierarchy
+//! - semantic_tokens: Semantic tokens for syntax highlighting
+//! - misc: Inlay hints, document links, code lens, and other features
 
+mod code_actions;
 mod completion;
 mod formatting;
+mod hierarchy;
 mod hover;
+mod misc;
 mod navigation;
 mod references;
+mod rename;
+mod semantic_tokens;
 mod symbols;
