+++ crates/perl-parser/src/lsp/server_impl/language/rename.rs
@@ -1,0 +1,239 @@
+//! Rename handlers for symbol renaming
+//!
+//! Handles textDocument/prepareRename and textDocument/rename requests.
+//! Supports both single-file and workspace-wide renaming.
+
+use super::super::*;
+
+impl LspServer {
+    /// Handle textDocument/prepareRename request
+    pub(crate) fn handle_prepare_rename(
+        &self,
+        params: Option<Value>,
+    ) -> Result<Option<Value>, JsonRpcError> {
+        if let Some(params) = params {
+            let uri = params["textDocument"]["uri"].as_str().unwrap_or("");
+            let line = params["position"]["line"].as_u64().unwrap_or(0) as u32;
+            let character = params["position"]["character"].as_u64().unwrap_or(0) as u32;
+
+            let documents = self.documents.lock().unwrap_or_else(|e| e.into_inner());
+            if let Some(doc) = self.get_document(&documents, uri) {
+                if let Some(_ast) = &doc.ast {
+                    let offset = self.pos16_to_offset(doc, line, character);
+
+                    // Get the token at the current position
+                    let token = self.get_token_at_position(&doc.text, offset);
+                    if !token.is_empty()
+                        && (token.starts_with('$')
+                            || token.starts_with('@')
+                            || token.starts_with('%')
+                            || token.chars().next().is_some_and(|c| c.is_alphabetic() || c == '_'))
+                    {
+                        // Find the token bounds
+                        let (start_offset, end_offset) = self.get_token_bounds(&doc.text, offset);
+                        let (start_line, start_char) = self.offset_to_pos16(doc, start_offset);
+                        let (end_line, end_char) = self.offset_to_pos16(doc, end_offset);
+
+                        // Return the range and placeholder text
+                        return Ok(Some(json!({
+                            "range": {
+                                "start": {
+                                    "line": start_line,
+                                    "character": start_char
+                                },
+                                "end": {
+                                    "line": end_line,
+                                    "character": end_char
+                                }
+                            },
+                            "placeholder": token
+                        })));
+                    }
+                }
+            }
+        }
+
+        // Return null if rename is not possible at this position
+        Ok(Some(json!(null)))
+    }
+
+    /// Handle textDocument/rename request (single file)
+    #[allow(dead_code)] // Dispatch uses handle_rename_workspace instead
+    pub(crate) fn handle_rename(
+        &self,
+        params: Option<Value>,
+    ) -> Result<Option<Value>, JsonRpcError> {
+        if let Some(params) = params {
+            let uri = params["textDocument"]["uri"].as_str().unwrap_or("");
+            let line = params["position"]["line"].as_u64().unwrap_or(0) as u32;
+            let character = params["position"]["character"].as_u64().unwrap_or(0) as u32;
+            let new_name = params["newName"].as_str().unwrap_or("");
+
+            // Validate the new name
+            if !self.is_valid_identifier(new_name) {
+                return Err(JsonRpcError {
+                    code: -32602,
+                    message: format!("Invalid identifier: {}", new_name),
+                    data: None,
+                });
+            }
+
+            let documents = self.documents.lock().unwrap_or_else(|e| e.into_inner());
+            if let Some(doc) = self.get_document(&documents, uri) {
+                if let Some(ref ast) = doc.ast {
+                    let offset = self.pos16_to_offset(doc, line, character);
+
+                    // Create semantic analyzer
+                    let analyzer = crate::semantic::SemanticAnalyzer::analyze(ast);
+
+                    // Find all references (including definition)
+                    let references = analyzer.find_all_references(offset, true);
+
+                    if !references.is_empty() {
+                        // Create text edits for all references
+                        let mut edits = Vec::new();
+                        for location in references {
+                            let (start_line, start_char) =
+                                self.offset_to_pos16(doc, location.start);
+                            let (end_line, end_char) = self.offset_to_pos16(doc, location.end);
+
+                            edits.push(json!({
+                                "range": {
+                                    "start": { "line": start_line, "character": start_char },
+                                    "end": { "line": end_line, "character": end_char }
+                                },
+                                "newText": new_name
+                            }));
+                        }
+
+                        // Return WorkspaceEdit
+                        return Ok(Some(json!({
+                            "changes": {
+                                uri: edits
+                            }
+                        })));
+                    }
+                }
+            }
+        }
+
+        // Return empty workspace edit if nothing to rename
+        Ok(Some(json!({
+            "changes": {}
+        })))
+    }
+
+    /// Handle textDocument/rename request with workspace support
+    pub(crate) fn handle_rename_workspace(
+        &self,
+        params: Option<Value>,
+    ) -> Result<Option<Value>, JsonRpcError> {
+        if let Some(p) = params {
+            if let (Some(uri), Some(line), Some(ch), Some(new_name)) = (
+                p.get("textDocument").and_then(|t| t.get("uri")).and_then(|s| s.as_str()),
+                p.get("position").and_then(|p| p.get("line")).and_then(|n| n.as_u64()),
+                p.get("position").and_then(|p| p.get("character")).and_then(|n| n.as_u64()),
+                p.get("newName").and_then(|s| s.as_str()),
+            ) {
+                let documents = self.documents.lock().unwrap_or_else(|e| e.into_inner());
+                if let Some(doc) = documents.get(uri) {
+                    if let Some(ref ast) = doc.ast {
+                        let offset = self.pos16_to_offset(doc, line as u32, ch as u32);
+                        let current_pkg = crate::declaration::current_package_at(ast, offset);
+                        if let Some(key) =
+                            crate::declaration::symbol_at_cursor(ast, offset, current_pkg)
+                        {
+                            #[cfg(feature = "workspace")]
+                            if let Some(ref idx) = self.workspace_index {
+                                let edits =
+                                    crate::workspace_rename::build_rename_edit(idx, &key, new_name);
+                                let ws_edit = crate::workspace_rename::to_workspace_edit(edits);
+                                return Ok(Some(ws_edit));
+                            }
+                        }
+                    }
+                }
+            }
+        }
+        // Return empty edit if we can't resolve
+        Ok(Some(json!({"changes": {}})))
+    }
+
+    /// Validate if a string is a valid Perl identifier
+    #[allow(dead_code)] // Used by handle_rename which is currently not dispatched
+    pub(crate) fn is_valid_identifier(&self, name: &str) -> bool {
+        if name.is_empty() {
+            return false;
+        }
+
+        let chars: Vec<char> = name.chars().collect();
+
+        // First character must be letter or underscore
+        if !chars[0].is_alphabetic() && chars[0] != '_' {
+            return false;
+        }
+
+        // Rest must be alphanumeric or underscore
+        for ch in &chars[1..] {
+            if !ch.is_alphanumeric() && *ch != '_' {
+                return false;
+            }
+        }
+
+        true
+    }
+
+    /// Get token at position (simple implementation)
+    pub(crate) fn get_token_at_position(&self, content: &str, offset: usize) -> String {
+        let chars: Vec<char> = content.chars().collect();
+        if offset >= chars.len() {
+            return String::new();
+        }
+
+        // Find word boundaries
+        let mut start = offset;
+        while start > 0
+            && (chars[start - 1].is_alphanumeric()
+                || chars[start - 1] == '_'
+                || chars[start - 1] == '$'
+                || chars[start - 1] == '@'
+                || chars[start - 1] == '%')
+        {
+            start -= 1;
+        }
+
+        let mut end = offset;
+        while end < chars.len() && (chars[end].is_alphanumeric() || chars[end] == '_') {
+            end += 1;
+        }
+
+        chars[start..end].iter().collect()
+    }
+
+    /// Get the bounds of the token at the given position
+    pub(crate) fn get_token_bounds(&self, content: &str, offset: usize) -> (usize, usize) {
+        let chars: Vec<char> = content.chars().collect();
+        if offset >= chars.len() {
+            return (offset, offset);
+        }
+
+        // Find word boundaries
+        let mut start = offset;
+        while start > 0
+            && (chars[start - 1].is_alphanumeric()
+                || chars[start - 1] == '_'
+                || chars[start - 1] == '$'
+                || chars[start - 1] == '@'
+                || chars[start - 1] == '%')
+        {
+            start -= 1;
+        }
+
+        let mut end = offset;
+        while end < chars.len() && (chars[end].is_alphanumeric() || chars[end] == '_') {
+            end += 1;
+        }
+
+        (start, end)
+    }
+}
