+++ crates/perl-parser/src/lsp/server_impl/mod.rs
@@ -35,7 +35,7 @@ use crate::{
     document_highlight::DocumentHighlightProvider,
     formatting::{CodeFormatter, FormattingOptions},
     implementation_provider::ImplementationProvider,
-    inlay_hints_provider::{InlayHintConfig, InlayHintsProvider},
+    // Note: InlayHintConfig and InlayHintsProvider are used in language/misc.rs
     // Import from new modular lsp structure
     // Note: JsonRpcError, JsonRpcRequest, JsonRpcResponse are pub use'd above
     lsp::protocol::{
@@ -274,1108 +274,94 @@ impl LspServer {
 
     // Note: completion handlers are implemented in language/completion.rs
 
-    /// Handle code action request
-    fn handle_code_action(&self, params: Option<Value>) -> Result<Option<Value>, JsonRpcError> {
-        if let Some(params) = params {
-            let uri = params["textDocument"]["uri"].as_str().unwrap_or("");
-            let start_line = params["range"]["start"]["line"].as_u64().unwrap_or(0) as u32;
-            let start_char = params["range"]["start"]["character"].as_u64().unwrap_or(0) as u32;
-            let end_line = params["range"]["end"]["line"].as_u64().unwrap_or(0) as u32;
-            let end_char = params["range"]["end"]["character"].as_u64().unwrap_or(0) as u32;
+    // Note: handle_code_action is implemented in language/code_actions.rs
 
-            let documents = self.documents.lock().unwrap();
-            if let Some(doc) = self.get_document(&documents, uri) {
-                if let Some(ast) = &doc.ast {
-                    let start_offset = self.pos16_to_offset(doc, start_line, start_char);
-                    let end_offset = self.pos16_to_offset(doc, end_line, end_char);
+    // Note: handle_prepare_type_hierarchy, handle_type_hierarchy_supertypes,
+    // handle_type_hierarchy_subtypes are implemented in language/hierarchy.rs
 
-                    // Get diagnostics from the document
-                    let diag_provider = DiagnosticsProvider::new(ast, doc.text.clone());
-                    let diagnostics =
-                        diag_provider.get_diagnostics(ast, &doc.parse_errors, &doc.text);
+    // Note: handle_prepare_rename, handle_rename, handle_rename_workspace are
+    // implemented in language/rename.rs
 
-                    // Get code actions from both providers
-                    let mut code_actions: Vec<Value> = Vec::new();
+    // Note: handle_code_actions_pragmas, handle_code_action_resolve are implemented
+    // in language/code_actions.rs
 
-                    // Add Perl::Critic quick fixes
-                    let builtin_analyzer = BuiltInAnalyzer::new();
-                    let violations = builtin_analyzer.analyze(ast, &doc.text);
-                    for violation in &violations {
-                        if let Some(quick_fix) =
-                            builtin_analyzer.get_quick_fix(violation, &doc.text)
-                        {
-                            let mut changes = HashMap::new();
-                            let (start_line, start_char) =
-                                self.offset_to_pos16(doc, violation.range.start.byte);
-                            let (end_line, end_char) =
-                                self.offset_to_pos16(doc, violation.range.end.byte);
+    // Note: handle_semantic_tokens is implemented in language/semantic_tokens.rs
 
-                            changes.insert(
-                                uri.to_string(),
-                                vec![json!({
-                                    "range": {
-                                        "start": {"line": start_line, "character": start_char},
-                                        "end": {"line": end_line, "character": end_char},
-                                    },
-                                    "newText": quick_fix.edit.new_text,
-                                })],
-                            );
+    // Note: handle_inlay_hints, handle_document_links, handle_selection_range, workspace_roots
+    // are implemented in language/misc.rs
 
-                            code_actions.push(json!({
-                                "title": quick_fix.title,
-                                "kind": "quickfix",
-                                "diagnostics": [{
-                                    "range": {
-                                        "start": {"line": start_line, "character": start_char},
-                                        "end": {"line": end_line, "character": end_char},
-                                    },
-                                    "severity": match violation.severity {
-                                        crate::perl_critic::Severity::Brutal |
-                                        crate::perl_critic::Severity::Cruel => 1, // Error
-                                        crate::perl_critic::Severity::Harsh => 2, // Warning
-                                        _ => 3, // Information
-                                    },
-                                    "code": violation.policy.clone(),
-                                    "source": "Perl::Critic",
-                                    "message": violation.description.clone()
-                                }],
-                                "edit": {
-                                    "changes": changes,
-                                },
-                            }));
-                        }
-                    }
+    // Note: is_valid_identifier, get_token_at_position, get_token_bounds are
+    // implemented in language/rename.rs
 
-                    // Get quick-fixes from the V2 provider (diagnostic-based)
-                    let provider_v2 = CodeActionsProviderV2::new(doc.text.clone());
-                    let quick_fixes =
-                        provider_v2.get_code_actions((start_offset, end_offset), &diagnostics);
+    /// Run a specific test
+    pub(crate) fn run_test(&self, test_id: &str) -> Result<Option<Value>, JsonRpcError> {
+        eprintln!("Running test: {}", test_id);
 
-                    for action in quick_fixes {
-                        let mut changes = HashMap::new();
-                        let (start_line, start_char) =
-                            self.offset_to_pos16(doc, action.edit.range.0);
-                        let (end_line, end_char) = self.offset_to_pos16(doc, action.edit.range.1);
-
-                        let edits = vec![json!({
-                            "range": {
-                                "start": {"line": start_line, "character": start_char},
-                                "end": {"line": end_line, "character": end_char},
-                            },
-                            "newText": action.edit.new_text,
-                        })];
-                        changes.insert(uri.to_string(), edits);
-
-                        code_actions.push(json!({
-                            "title": action.title,
-                            "kind": match action.kind {
-                                InternalCodeActionKindV2::QuickFix => "quickfix",
-                                InternalCodeActionKindV2::Refactor => "refactor",
-                                InternalCodeActionKindV2::RefactorExtract => "refactor.extract",
-                                InternalCodeActionKindV2::RefactorInline => "refactor.inline",
-                                InternalCodeActionKindV2::RefactorRewrite => "refactor.rewrite",
-                            },
-                            "edit": {
-                                "changes": changes,
-                            },
-                        }));
-                    }
-
-                    // Get refactorings from the original provider (AST-based)
-                    let provider = CodeActionsProvider::new(doc.text.clone());
-                    let actions =
-                        provider.get_code_actions(ast, (start_offset, end_offset), &diagnostics);
-
-                    for action in actions {
-                        let mut changes = HashMap::new();
-                        let edits: Vec<Value> = action
-                            .edit
-                            .changes
-                            .into_iter()
-                            .map(|edit| {
-                                let (start_line, start_char) =
-                                    self.offset_to_pos16(doc, edit.location.start);
-                                let (end_line, end_char) =
-                                    self.offset_to_pos16(doc, edit.location.end);
-                                json!({
-                                    "range": {
-                                        "start": {"line": start_line, "character": start_char},
-                                        "end": {"line": end_line, "character": end_char},
-                                    },
-                                    "newText": edit.new_text,
-                                })
-                            })
-                            .collect();
-                        changes.insert(uri.to_string(), edits);
-
-                        code_actions.push(json!({
-                            "title": action.title,
-                            "kind": match action.kind {
-                                InternalCodeActionKind::QuickFix => "quickfix",
-                                InternalCodeActionKind::Refactor => "refactor",
-                                InternalCodeActionKind::RefactorExtract => "refactor.extract",
-                                InternalCodeActionKind::RefactorInline => "refactor.inline",
-                                InternalCodeActionKind::RefactorRewrite => "refactor.rewrite",
-                                InternalCodeActionKind::Source => "source",
-                                InternalCodeActionKind::SourceOrganizeImports => "source.organizeImports",
-                                InternalCodeActionKind::SourceFixAll => "source.fixAll",
-                            },
-                            "edit": {
-                                "changes": changes,
-                            },
-                        }));
-                    }
-
-                    // Get enhanced refactorings (extract variable, convert loops, etc.)
-                    let enhanced_provider = EnhancedCodeActionsProvider::new(doc.text.clone());
-                    let enhanced_actions = enhanced_provider
-                        .get_enhanced_refactoring_actions(ast, (start_offset, end_offset));
-
-                    // Add test generation actions
-                    let test_generator = TestGenerator::new("Test::More");
-                    let subroutines = test_generator.find_subroutines(ast);
-
-                    for action in enhanced_actions {
-                        let mut changes = HashMap::new();
-                        let edits: Vec<Value> = action
-                            .edit
-                            .changes
-                            .into_iter()
-                            .map(|edit| {
-                                let (start_line, start_char) =
-                                    self.offset_to_pos16(doc, edit.location.start);
-                                let (end_line, end_char) =
-                                    self.offset_to_pos16(doc, edit.location.end);
-                                json!({
-                                    "range": {
-                                        "start": {"line": start_line, "character": start_char},
-                                        "end": {"line": end_line, "character": end_char},
-                                    },
-                                    "newText": edit.new_text,
-                                })
-                            })
-                            .collect();
-                        changes.insert(uri.to_string(), edits);
-
-                        code_actions.push(json!({
-                            "title": action.title,
-                            "kind": match action.kind {
-                                InternalCodeActionKind::QuickFix => "quickfix",
-                                InternalCodeActionKind::Refactor => "refactor",
-                                InternalCodeActionKind::RefactorExtract => "refactor.extract",
-                                InternalCodeActionKind::RefactorInline => "refactor.inline",
-                                InternalCodeActionKind::RefactorRewrite => "refactor.rewrite",
-                                InternalCodeActionKind::Source => "source",
-                                InternalCodeActionKind::SourceOrganizeImports => "source.organizeImports",
-                                InternalCodeActionKind::SourceFixAll => "source.fixAll",
-                            },
-                            "edit": {
-                                "changes": changes,
-                            },
-                        }));
-                    }
-
-                    // Add test generation actions for subroutines in range
-                    for sub_info in subroutines {
-                        // Check if cursor is near this subroutine
-                        let test_code =
-                            test_generator.generate_test(&sub_info.name, sub_info.param_count);
-                        code_actions.push(json!({
-                            "title": format!("Generate test for '{}'", sub_info.name),
-                            "kind": "source",
-                            "command": {
-                                "title": "Generate test",
-                                "command": "perl.generateTest",
-                                "arguments": [json!({
-                                    "uri": uri,
-                                    "name": sub_info.name,
-                                    "test": test_code
-                                })]
-                            }
-                        }));
-                    }
-
-                    // Always offer generic debug actions when there are diagnostics
-                    if !diagnostics.is_empty() {
-                        // Add debug print action
-                        code_actions.push(json!({
-                            "title": "Add debug print",
-                            "kind": "refactor.rewrite",
-                            "command": {
-                                "title": "Add debug print",
-                                "command": "perl.addDebugPrint",
-                                "arguments": [json!({ "uri": uri, "range": {
-                                    "start": {"line": start_line, "character": start_char},
-                                    "end": {"line": end_line, "character": end_char}
-                                }})]
-                            }
-                        }));
-
-                        // Extract variable action
-                        code_actions.push(json!({
-                            "title": "Extract variable",
-                            "kind": "refactor.extract",
-                            "command": {
-                                "title": "Extract variable",
-                                "command": "perl.extractVariable",
-                                "arguments": [json!({ "uri": uri, "range": {
-                                    "start": {"line": start_line, "character": start_char},
-                                    "end": {"line": end_line, "character": end_char}
-                                }})]
-                            }
-                        }));
-                    }
-
-                    return Ok(Some(json!(code_actions)));
-                } else {
-                    // No AST (parse error), but we can still offer some actions
-                    let mut code_actions: Vec<Value> = Vec::new();
-
-                    // Check if source lacks strict/warnings
-                    if !doc.text.contains("use strict") || !doc.text.contains("use warnings") {
-                        let mut changes = HashMap::new();
-                        // Find first non-shebang line
-                        let insert_pos = if doc.text.starts_with("#!") {
-                            doc.text.find('\n').map(|p| p + 1).unwrap_or(0)
-                        } else {
-                            0
-                        };
-
-                        let new_text = if !doc.text.contains("use strict")
-                            && !doc.text.contains("use warnings")
-                        {
-                            "use strict;\nuse warnings;\n\n"
-                        } else if !doc.text.contains("use strict") {
-                            "use strict;\n"
-                        } else {
-                            "use warnings;\n"
-                        };
-
-                        let (line, char) = self.offset_to_pos16(doc, insert_pos);
-                        changes.insert(
-                            uri.to_string(),
-                            vec![json!({
-                                "range": {
-                                    "start": {"line": line, "character": char},
-                                    "end": {"line": line, "character": char},
-                                },
-                                "newText": new_text,
-                            })],
-                        );
-
-                        code_actions.push(json!({
-                            "title": "Add 'use strict' and 'use warnings'",
-                            "kind": "quickfix",
-                            "edit": {
-                                "changes": changes,
-                            },
-                        }));
-                    }
-
-                    // Always offer debug actions for files with issues
-                    code_actions.push(json!({
-                        "title": "Add debug print",
-                        "kind": "refactor.rewrite",
-                        "command": {
-                            "title": "Add debug print",
-                            "command": "perl.addDebugPrint",
-                            "arguments": [json!({ "uri": uri })]
-                        }
-                    }));
-
-                    // Check for global variables that could use 'my' declarations
-                    let global_var_pattern =
-                        regex::Regex::new(r"(?m)^(\$|\@|\%)[a-zA-Z_]\w*\s*=").ok();
-                    if let Some(re) = global_var_pattern {
-                        if re.is_match(&doc.text) {
-                            code_actions.push(json!({
-                                "title": "Convert globals to 'my' declarations",
-                                "kind": "refactor.rewrite",
-                                "command": {
-                                    "title": "Convert to my declarations",
-                                    "command": "perl.convertToMyDeclarations",
-                                    "arguments": [json!({ "uri": uri })]
-                                }
-                            }));
-                        }
-                    }
-
-                    return Ok(Some(json!(code_actions)));
-                }
-            }
+        // Parse test ID to get URI and test name
+        let parts: Vec<&str> = test_id.split("::").collect();
+        if parts.len() < 2 {
+            return Ok(Some(json!({"status": "error", "message": "Invalid test ID"})));
         }
 
-        Ok(Some(json!([])))
+        let uri = parts[0];
+        let test_name = parts[1..].join("::");
+
+        let documents = self.documents.lock().unwrap();
+        if let Some(doc) = documents.get(uri) {
+            let runner = TestRunner::new(doc.text.clone(), uri.to_string());
+            let results = runner.run_test(&test_name);
+
+            // Convert results to JSON
+            let json_results: Vec<Value> = results
+                .into_iter()
+                .map(|result| {
+                    json!({
+                        "testId": result.test_id,
+                        "status": result.status.as_str(),
+                        "message": result.message,
+                        "duration": result.duration
+                    })
+                })
+                .collect();
+
+            return Ok(Some(json!({
+                "status": "success",
+                "results": json_results
+            })));
+        }
+
+        Ok(Some(document_not_found_error()))
     }
 
-    /// Handle textDocument/prepareTypeHierarchy request
-    fn handle_prepare_type_hierarchy(
-        &self,
-        params: Option<Value>,
-    ) -> Result<Option<Value>, JsonRpcError> {
-        if let Some(params) = params {
-            let uri = params["textDocument"]["uri"].as_str().unwrap_or("");
-            let line = params["position"]["line"].as_u64().unwrap_or(0) as u32;
-            let character = params["position"]["character"].as_u64().unwrap_or(0) as u32;
+    /// Run all tests in a file
+    pub(crate) fn run_test_file(&self, uri: &str) -> Result<Option<Value>, JsonRpcError> {
+        eprintln!("Running test file: {}", uri);
 
-            let documents = self.documents.lock().unwrap();
-            if let Some(doc) = self.get_document(&documents, uri) {
-                let offset = self.pos16_to_offset(doc, line, character);
+        let documents = self.documents.lock().unwrap();
+        if let Some(doc) = documents.get(uri) {
+            let runner = TestRunner::new(doc.text.clone(), uri.to_string());
+            let results = runner.run_test(uri);
 
-                // Try AST-based approach first
-                if let Some(ref ast) = doc.ast {
-                    // Create type hierarchy provider
-                    let provider = TypeHierarchyProvider::new();
+            // Convert results to JSON
+            let json_results: Vec<Value> = results
+                .into_iter()
+                .map(|result| {
+                    json!({
+                        "testId": result.test_id,
+                        "status": result.status.as_str(),
+                        "message": result.message,
+                        "duration": result.duration
+                    })
+                })
+                .collect();
 
-                    // Prepare type hierarchy at the position
-                    if let Some(items) = provider.prepare(ast, &doc.text, offset) {
-                        let lsp_items: Vec<Value> = items
-                            .iter()
-                            .map(|item| {
-                                json!({
-                                    "name": item.name,
-                                    "kind": item.kind as u32,
-                                    "uri": uri,
-                                    "range": {
-                                        "start": {
-                                            "line": item.range.start.line,
-                                            "character": item.range.start.character,
-                                        },
-                                        "end": {
-                                            "line": item.range.end.line,
-                                            "character": item.range.end.character,
-                                        },
-                                    },
-                                    "selectionRange": {
-                                        "start": {
-                                            "line": item.selection_range.start.line,
-                                            "character": item.selection_range.start.character,
-                                        },
-                                        "end": {
-                                            "line": item.selection_range.end.line,
-                                            "character": item.selection_range.end.character,
-                                        },
-                                    },
-                                    "detail": item.detail,
-                                    "data": {
-                                        "uri": uri,
-                                        "name": item.name,
-                                    },
-                                })
-                            })
-                            .collect();
-
-                        return Ok(Some(json!(lsp_items)));
-                    }
-                }
-
-                // Fallback to regex-based approach
-                let sub_regex = regex::Regex::new(r"\bsub\s+([a-zA-Z_]\w*)\b").unwrap();
-                let package_regex = regex::Regex::new(r"\bpackage\s+([a-zA-Z_][\w:]*)\b").unwrap();
-
-                // Find all subs and packages with their positions
-                let mut exact_sub: Option<(String, usize, usize)> = None;
-                for cap in sub_regex.captures_iter(&doc.text) {
-                    if let (Some(m), Some(name)) = (cap.get(0), cap.get(1)) {
-                        if offset >= m.start() && offset <= m.end() {
-                            // Exact match - cursor is on this sub
-                            exact_sub = Some((name.as_str().to_string(), m.start(), m.end()));
-                            break;
-                        }
-                    }
-                }
-
-                if let Some((name, start, end)) = exact_sub {
-                    let start_pos = doc.line_starts.offset_to_position_rope(&doc.rope, start);
-                    let end_pos = doc.line_starts.offset_to_position_rope(&doc.rope, end);
-                    return Ok(Some(json!([{
-                        "name": name,
-                        "kind": 12, // Function
-                        "uri": uri,
-                        "range": {
-                            "start": { "line": start_pos.0, "character": start_pos.1 },
-                            "end": { "line": end_pos.0, "character": end_pos.1 },
-                        },
-                        "selectionRange": {
-                            "start": { "line": start_pos.0, "character": start_pos.1 },
-                            "end": { "line": end_pos.0, "character": end_pos.1 },
-                        },
-                        "detail": "sub",
-                        "data": { "uri": uri, "name": name },
-                    }])));
-                }
-
-                // Check packages
-                let mut exact_pkg: Option<(String, usize, usize)> = None;
-                for cap in package_regex.captures_iter(&doc.text) {
-                    if let (Some(m), Some(name)) = (cap.get(0), cap.get(1)) {
-                        if offset >= m.start() && offset <= m.end() {
-                            // Exact match - cursor is on this package
-                            exact_pkg = Some((name.as_str().to_string(), m.start(), m.end()));
-                            break;
-                        }
-                    }
-                }
-
-                if let Some((name, start, end)) = exact_pkg {
-                    let start_pos = doc.line_starts.offset_to_position_rope(&doc.rope, start);
-                    let end_pos = doc.line_starts.offset_to_position_rope(&doc.rope, end);
-                    return Ok(Some(json!([{
-                        "name": name,
-                        "kind": 5, // Class
-                        "uri": uri,
-                        "range": {
-                            "start": { "line": start_pos.0, "character": start_pos.1 },
-                            "end": { "line": end_pos.0, "character": end_pos.1 },
-                        },
-                        "selectionRange": {
-                            "start": { "line": start_pos.0, "character": start_pos.1 },
-                            "end": { "line": end_pos.0, "character": end_pos.1 },
-                        },
-                        "detail": "package",
-                        "data": { "uri": uri, "name": name },
-                    }])));
-                }
-            }
+            return Ok(Some(json!({
+                "status": "success",
+                "results": json_results
+            })));
         }
 
-        Ok(Some(json!([])))
-    }
-
-    /// Handle typeHierarchy/supertypes request
-    fn handle_type_hierarchy_supertypes(
-        &self,
-        params: Option<Value>,
-    ) -> Result<Option<Value>, JsonRpcError> {
-        if let Some(params) = params {
-            if let Some(item) = params.get("item") {
-                let uri = item["data"]["uri"].as_str().unwrap_or("");
-                let name = item["data"]["name"].as_str().unwrap_or("");
-
-                let documents = self.documents.lock().unwrap();
-                if let Some(doc) = documents.get(uri) {
-                    if let Some(ref ast) = doc.ast {
-                        // Create type hierarchy provider
-                        let provider = TypeHierarchyProvider::new();
-
-                        // Create item from request
-                        let type_item = crate::type_hierarchy::TypeHierarchyItem {
-                            name: name.to_string(),
-                            kind: crate::type_hierarchy::SymbolKind::Class,
-                            uri: uri.to_string(),
-                            range: crate::type_hierarchy::Range {
-                                start: crate::type_hierarchy::Position { line: 0, character: 0 },
-                                end: crate::type_hierarchy::Position { line: 0, character: 0 },
-                            },
-                            selection_range: crate::type_hierarchy::Range {
-                                start: crate::type_hierarchy::Position { line: 0, character: 0 },
-                                end: crate::type_hierarchy::Position { line: 0, character: 0 },
-                            },
-                            detail: None,
-                            data: None,
-                        };
-
-                        // Find supertypes
-                        let supertypes = provider.find_supertypes(ast, &type_item);
-
-                        let lsp_items: Vec<Value> = supertypes
-                            .iter()
-                            .map(|item| {
-                                json!({
-                                    "name": item.name,
-                                    "kind": item.kind as u32,
-                                    "uri": uri,
-                                    "range": {
-                                        "start": {
-                                            "line": item.range.start.line,
-                                            "character": item.range.start.character,
-                                        },
-                                        "end": {
-                                            "line": item.range.end.line,
-                                            "character": item.range.end.character,
-                                        },
-                                    },
-                                    "selectionRange": {
-                                        "start": {
-                                            "line": item.selection_range.start.line,
-                                            "character": item.selection_range.start.character,
-                                        },
-                                        "end": {
-                                            "line": item.selection_range.end.line,
-                                            "character": item.selection_range.end.character,
-                                        },
-                                    },
-                                    "detail": item.detail,
-                                    "data": {
-                                        "uri": uri,
-                                        "name": item.name,
-                                    },
-                                })
-                            })
-                            .collect();
-
-                        return Ok(Some(json!(lsp_items)));
-                    }
-                }
-            }
-        }
-
-        Ok(Some(json!([])))
-    }
-
-    /// Handle typeHierarchy/subtypes request
-    fn handle_type_hierarchy_subtypes(
-        &self,
-        params: Option<Value>,
-    ) -> Result<Option<Value>, JsonRpcError> {
-        if let Some(params) = params {
-            if let Some(item) = params.get("item") {
-                let uri = item["data"]["uri"].as_str().unwrap_or("");
-                let name = item["data"]["name"].as_str().unwrap_or("");
-
-                let documents = self.documents.lock().unwrap();
-                if let Some(doc) = documents.get(uri) {
-                    if let Some(ref ast) = doc.ast {
-                        // Create type hierarchy provider
-                        let provider = TypeHierarchyProvider::new();
-
-                        // Create item from request
-                        let type_item = crate::type_hierarchy::TypeHierarchyItem {
-                            name: name.to_string(),
-                            kind: crate::type_hierarchy::SymbolKind::Class,
-                            uri: uri.to_string(),
-                            range: crate::type_hierarchy::Range {
-                                start: crate::type_hierarchy::Position { line: 0, character: 0 },
-                                end: crate::type_hierarchy::Position { line: 0, character: 0 },
-                            },
-                            selection_range: crate::type_hierarchy::Range {
-                                start: crate::type_hierarchy::Position { line: 0, character: 0 },
-                                end: crate::type_hierarchy::Position { line: 0, character: 0 },
-                            },
-                            detail: None,
-                            data: None,
-                        };
-
-                        // Find subtypes
-                        let subtypes = provider.find_subtypes(ast, &type_item);
-
-                        let lsp_items: Vec<Value> = subtypes
-                            .iter()
-                            .map(|item| {
-                                json!({
-                                    "name": item.name,
-                                    "kind": item.kind as u32,
-                                    "uri": uri,
-                                    "range": {
-                                        "start": {
-                                            "line": item.range.start.line,
-                                            "character": item.range.start.character,
-                                        },
-                                        "end": {
-                                            "line": item.range.end.line,
-                                            "character": item.range.end.character,
-                                        },
-                                    },
-                                    "selectionRange": {
-                                        "start": {
-                                            "line": item.selection_range.start.line,
-                                            "character": item.selection_range.start.character,
-                                        },
-                                        "end": {
-                                            "line": item.selection_range.end.line,
-                                            "character": item.selection_range.end.character,
-                                        },
-                                    },
-                                    "detail": item.detail,
-                                    "data": {
-                                        "uri": uri,
-                                        "name": item.name,
-                                    },
-                                })
-                            })
-                            .collect();
-
-                        return Ok(Some(json!(lsp_items)));
-                    }
-                }
-            }
-        }
-
-        Ok(Some(json!([])))
-    }
-
-    /// Handle textDocument/rename request
-    fn handle_prepare_rename(&self, params: Option<Value>) -> Result<Option<Value>, JsonRpcError> {
-        if let Some(params) = params {
-            let uri = params["textDocument"]["uri"].as_str().unwrap_or("");
-            let line = params["position"]["line"].as_u64().unwrap_or(0) as u32;
-            let character = params["position"]["character"].as_u64().unwrap_or(0) as u32;
-
-            let documents = self.documents.lock().unwrap();
-            if let Some(doc) = self.get_document(&documents, uri) {
-                if let Some(_ast) = &doc.ast {
-                    let offset = self.pos16_to_offset(doc, line, character);
-
-                    // Get the token at the current position
-                    let token = self.get_token_at_position(&doc.text, offset);
-                    if !token.is_empty()
-                        && (token.starts_with('$')
-                            || token.starts_with('@')
-                            || token.starts_with('%')
-                            || token.chars().next().is_some_and(|c| c.is_alphabetic() || c == '_'))
-                    {
-                        // Find the token bounds
-                        let (start_offset, end_offset) = self.get_token_bounds(&doc.text, offset);
-                        let (start_line, start_char) = self.offset_to_pos16(doc, start_offset);
-                        let (end_line, end_char) = self.offset_to_pos16(doc, end_offset);
-
-                        // Return the range and placeholder text
-                        return Ok(Some(json!({
-                            "range": {
-                                "start": {
-                                    "line": start_line,
-                                    "character": start_char
-                                },
-                                "end": {
-                                    "line": end_line,
-                                    "character": end_char
-                                }
-                            },
-                            "placeholder": token
-                        })));
-                    }
-                }
-            }
-        }
-
-        // Return null if rename is not possible at this position
-        Ok(Some(json!(null)))
-    }
-
-    fn handle_rename(&self, params: Option<Value>) -> Result<Option<Value>, JsonRpcError> {
-        if let Some(params) = params {
-            let uri = params["textDocument"]["uri"].as_str().unwrap_or("");
-            let line = params["position"]["line"].as_u64().unwrap_or(0) as u32;
-            let character = params["position"]["character"].as_u64().unwrap_or(0) as u32;
-            let new_name = params["newName"].as_str().unwrap_or("");
-
-            // Validate the new name
-            if !self.is_valid_identifier(new_name) {
-                return Err(JsonRpcError {
-                    code: -32602,
-                    message: format!("Invalid identifier: {}", new_name),
-                    data: None,
-                });
-            }
-
-            let documents = self.documents.lock().unwrap();
-            if let Some(doc) = self.get_document(&documents, uri) {
-                if let Some(ref ast) = doc.ast {
-                    let offset = self.pos16_to_offset(doc, line, character);
-
-                    // Create semantic analyzer
-                    let analyzer = crate::semantic::SemanticAnalyzer::analyze(ast);
-
-                    // Find all references (including definition)
-                    let references = analyzer.find_all_references(offset, true);
-
-                    if !references.is_empty() {
-                        // Create text edits for all references
-                        let mut edits = Vec::new();
-                        for location in references {
-                            let (start_line, start_char) =
-                                self.offset_to_pos16(doc, location.start);
-                            let (end_line, end_char) = self.offset_to_pos16(doc, location.end);
-
-                            edits.push(json!({
-                                "range": {
-                                    "start": { "line": start_line, "character": start_char },
-                                    "end": { "line": end_line, "character": end_char }
-                                },
-                                "newText": new_name
-                            }));
-                        }
-
-                        // Return WorkspaceEdit
-                        return Ok(Some(json!({
-                            "changes": {
-                                uri: edits
-                            }
-                        })));
-                    }
-                }
-            }
-        }
-
-        // Return empty workspace edit if nothing to rename
-        Ok(Some(json!({
-            "changes": {}
-        })))
-    }
-
-    /// Handle textDocument/rename request with workspace support
-    fn handle_rename_workspace(
-        &self,
-        params: Option<Value>,
-    ) -> Result<Option<Value>, JsonRpcError> {
-        if let Some(p) = params {
-            if let (Some(uri), Some(line), Some(ch), Some(new_name)) = (
-                p.get("textDocument").and_then(|t| t.get("uri")).and_then(|s| s.as_str()),
-                p.get("position").and_then(|p| p.get("line")).and_then(|n| n.as_u64()),
-                p.get("position").and_then(|p| p.get("character")).and_then(|n| n.as_u64()),
-                p.get("newName").and_then(|s| s.as_str()),
-            ) {
-                let documents = self.documents.lock().unwrap();
-                if let Some(doc) = documents.get(uri) {
-                    if let Some(ref ast) = doc.ast {
-                        let offset = self.pos16_to_offset(doc, line as u32, ch as u32);
-                        let current_pkg = crate::declaration::current_package_at(ast, offset);
-                        if let Some(key) =
-                            crate::declaration::symbol_at_cursor(ast, offset, current_pkg)
-                        {
-                            #[cfg(feature = "workspace")]
-                            if let Some(ref idx) = self.workspace_index {
-                                let edits =
-                                    crate::workspace_rename::build_rename_edit(idx, &key, new_name);
-                                let ws_edit = crate::workspace_rename::to_workspace_edit(edits);
-                                return Ok(Some(ws_edit));
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        // Return empty edit if we can't resolve
-        Ok(Some(json!({"changes": {}})))
-    }
-
-    /// Handle textDocument/codeAction request for pragmas
-    fn handle_code_actions_pragmas(
-        &self,
-        params: Option<Value>,
-    ) -> Result<Option<Value>, JsonRpcError> {
-        if let Some(p) = params {
-            if let Some(uri) = p["textDocument"]["uri"].as_str() {
-                let documents = self.documents.lock().unwrap();
-                if let Some(doc) = documents.get(uri) {
-                    let mut actions =
-                        crate::code_actions_pragmas::missing_pragmas_actions(uri, &doc.text);
-
-                    // Fill in edits with proper ranges
-                    for a in &mut actions {
-                        let data_info = (
-                            a.get("data")
-                                .and_then(|d| d.get("uri"))
-                                .and_then(|s| s.as_str())
-                                .map(|s| s.to_string()),
-                            a.get("data").and_then(|d| d.get("insertAt")).and_then(|n| n.as_u64()),
-                            a.get("data")
-                                .and_then(|d| d.get("text"))
-                                .and_then(|s| s.as_str())
-                                .map(|s| s.to_string()),
-                        );
-
-                        if let (Some(u), Some(off), Some(txt)) = data_info {
-                            if let Some(obj) = a.as_object_mut() {
-                                let edit_range = if off as usize >= doc.text.len() {
-                                    let end = self.get_document_end_position(&doc.text);
-                                    json!({"start": end.clone(), "end": end })
-                                } else {
-                                    let (line, col) = self.offset_to_pos16(doc, off as usize);
-                                    json!({
-                                        "start": {"line": line, "character": col},
-                                        "end": {"line": line, "character": col}
-                                    })
-                                };
-
-                                obj.insert(
-                                    "edit".into(),
-                                    json!({
-                                        "changes": {
-                                            u: [{
-                                                "range": edit_range,
-                                                "newText": txt
-                                            }]
-                                        }
-                                    }),
-                                );
-                                obj.remove("data");
-                            }
-                        }
-                    }
-                    return Ok(Some(json!(actions)));
-                }
-            }
-        }
-        Ok(Some(json!([])))
-    }
-
-    /// Handle codeAction/resolve request
-    fn handle_code_action_resolve(
-        &self,
-        params: Option<Value>,
-    ) -> Result<Option<Value>, JsonRpcError> {
-        if let Some(mut action) = params {
-            // The action should already have minimal information
-            // We now need to compute the actual edits
-
-            if let Some(kind) = action.get("kind").and_then(|k| k.as_str()) {
-                if kind == "quickfix" {
-                    // For quickfix actions, compute the workspace edit now
-                    if let Some(data) = action.get("data") {
-                        if let Some(uri) = data.get("uri").and_then(|u| u.as_str()) {
-                            let documents = self.documents.lock().unwrap();
-                            if self.get_document(&documents, uri).is_some() {
-                                // Example: Add "use strict;" at the beginning
-                                if let Some(pragma) = data.get("pragma").and_then(|p| p.as_str()) {
-                                    let text = format!("{}\n", pragma);
-                                    let edit = json!({
-                                        "changes": {
-                                            uri: [{
-                                                "range": {
-                                                    "start": {"line": 0, "character": 0},
-                                                    "end": {"line": 0, "character": 0}
-                                                },
-                                                "newText": text
-                                            }]
-                                        }
-                                    });
-
-                                    if let Some(obj) = action.as_object_mut() {
-                                        obj.insert("edit".to_string(), edit);
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-
-            Ok(Some(action))
-        } else {
-            Ok(None)
-        }
-    }
-
-    /// Handle textDocument/semanticTokens/full request  
-    fn handle_semantic_tokens(&self, params: Option<Value>) -> Result<Option<Value>, JsonRpcError> {
-        if let Some(p) = params {
-            let uri = p["textDocument"]["uri"].as_str().ok_or_else(|| JsonRpcError {
-                code: INVALID_PARAMS,
-                message: "Missing textDocument.uri".into(),
-                data: None,
-            })?;
-            let documents = self.documents.lock().unwrap();
-            let doc = self.get_document(&documents, uri).ok_or_else(|| JsonRpcError {
-                code: INVALID_REQUEST,
-                message: format!("Document not open: {}", uri),
-                data: None,
-            })?;
-            if let Some(ref ast) = doc.ast {
-                let data =
-                    crate::semantic_tokens::collect_semantic_tokens(ast, &doc.text, &|off| {
-                        self.offset_to_pos16(doc, off)
-                    });
-                return Ok(Some(json!({ "data": data.into_iter().flatten().collect::<Vec<_>>() })));
-            }
-        }
-        Ok(Some(json!({ "data": [] })))
-    }
-
-    /// Handle textDocument/inlayHint request
-    fn handle_inlay_hints(&self, params: Option<Value>) -> Result<Option<Value>, JsonRpcError> {
-        if let Some(p) = params {
-            let uri = p["textDocument"]["uri"].as_str().ok_or_else(|| JsonRpcError {
-                code: INVALID_PARAMS,
-                message: "Missing textDocument.uri".into(),
-                data: None,
-            })?;
-
-            // Extract the range parameter (required by LSP spec)
-            let range = if let Some(range_val) = p.get("range") {
-                let start_line = range_val["start"]["line"].as_u64().unwrap_or(0) as u32;
-                let start_char = range_val["start"]["character"].as_u64().unwrap_or(0) as u32;
-                let end_line = range_val["end"]["line"].as_u64().unwrap_or(u32::MAX as u64) as u32;
-                let end_char =
-                    range_val["end"]["character"].as_u64().unwrap_or(u32::MAX as u64) as u32;
-                Some(crate::positions::Range::new(start_line, start_char, end_line, end_char))
-            } else {
-                None
-            };
-
-            let documents = self.documents.lock().unwrap();
-            let doc = self.get_document(&documents, uri).ok_or_else(|| JsonRpcError {
-                code: INVALID_REQUEST,
-                message: format!("Document not open: {}", uri),
-                data: None,
-            })?;
-            if let Some(ref ast) = doc.ast {
-                let mut hints = Vec::new();
-                hints.extend(crate::inlay_hints::parameter_hints(
-                    ast,
-                    &|off| self.offset_to_pos16(doc, off),
-                    range,
-                ));
-                hints.extend(crate::inlay_hints::trivial_type_hints(
-                    ast,
-                    &|off| self.offset_to_pos16(doc, off),
-                    range,
-                ));
-                return Ok(Some(json!(hints)));
-            }
-        }
-        Ok(Some(json!([])))
-    }
-
-    /// Handle textDocument/documentLink request
-    fn handle_document_links(&self, params: Option<Value>) -> Result<Option<Value>, JsonRpcError> {
-        if let Some(p) = params {
-            let uri = p["textDocument"]["uri"].as_str().ok_or_else(|| JsonRpcError {
-                code: INVALID_PARAMS,
-                message: "Missing textDocument.uri".into(),
-                data: None,
-            })?;
-            let documents = self.documents.lock().unwrap();
-            let doc = self.get_document(&documents, uri).ok_or_else(|| JsonRpcError {
-                code: INVALID_REQUEST,
-                message: format!("Document not open: {}", uri),
-                data: None,
-            })?;
-
-            // Get workspace roots from initialization params
-            let roots = self.workspace_roots();
-            let links = crate::document_links::compute_links(uri, &doc.text, &roots);
-            Ok(Some(json!(links)))
-        } else {
-            Ok(Some(json!([])))
-        }
-    }
-
-    /// Handle textDocument/selectionRange request
-    fn handle_selection_range(&self, params: Option<Value>) -> Result<Option<Value>, JsonRpcError> {
-        if let Some(p) = params {
-            let uri = p["textDocument"]["uri"].as_str().ok_or_else(|| JsonRpcError {
-                code: INVALID_PARAMS,
-                message: "Missing textDocument.uri".into(),
-                data: None,
-            })?;
-            let positions = p["positions"].as_array().ok_or_else(|| JsonRpcError {
-                code: INVALID_PARAMS,
-                message: "Missing positions array".into(),
-                data: None,
-            })?;
-
-            let documents = self.documents.lock().unwrap();
-            let doc = self.get_document(&documents, uri).ok_or_else(|| JsonRpcError {
-                code: INVALID_REQUEST,
-                message: format!("Document not open: {}", uri),
-                data: None,
-            })?;
-
-            let mut out = Vec::new();
-            if let Some(ref ast) = doc.ast {
-                // Build parent map if not cached
-                let parent_map = crate::selection_range::build_parent_map(ast);
-
-                for pos in positions {
-                    let line = pos["line"].as_u64().unwrap_or(0) as u32;
-                    let col = pos["character"].as_u64().unwrap_or(0) as u32;
-                    let off = self.pos16_to_offset(doc, line, col);
-                    let chain =
-                        crate::selection_range::selection_chain(ast, &parent_map, off, &|o| {
-                            self.offset_to_pos16(doc, o)
-                        });
-                    out.push(chain);
-                }
-            }
-            Ok(Some(json!(out)))
-        } else {
-            Ok(Some(json!([])))
-        }
-    }
-
-    /// Get workspace roots from initialization
-    fn workspace_roots(&self) -> Vec<url::Url> {
-        // In a real implementation, store these from initialize params
-        // For now, return empty vec
-        vec![]
-    }
-
-    /// Validate if a string is a valid Perl identifier
-    fn is_valid_identifier(&self, name: &str) -> bool {
-        if name.is_empty() {
-            return false;
-        }
-
-        let chars: Vec<char> = name.chars().collect();
-
-        // First character must be letter or underscore
-        if !chars[0].is_alphabetic() && chars[0] != '_' {
-            return false;
-        }
-
-        // Rest must be alphanumeric or underscore
-        for ch in &chars[1..] {
-            if !ch.is_alphanumeric() && *ch != '_' {
-                return false;
-            }
-        }
-
-        true
-    }
-
-    /// Get token at position (simple implementation)
-    fn get_token_at_position(&self, content: &str, offset: usize) -> String {
-        let chars: Vec<char> = content.chars().collect();
-        if offset >= chars.len() {
-            return String::new();
-        }
-
-        // Find word boundaries
-        let mut start = offset;
-        while start > 0
-            && (chars[start - 1].is_alphanumeric()
-                || chars[start - 1] == '_'
-                || chars[start - 1] == '$'
-                || chars[start - 1] == '@'
-                || chars[start - 1] == '%')
-        {
-            start -= 1;
-        }
-
-        let mut end = offset;
-        while end < chars.len() && (chars[end].is_alphanumeric() || chars[end] == '_') {
-            end += 1;
-        }
-
-        chars[start..end].iter().collect()
-    }
-
-    /// Get the bounds of the token at the given position
-    fn get_token_bounds(&self, content: &str, offset: usize) -> (usize, usize) {
-        let chars: Vec<char> = content.chars().collect();
-        if offset >= chars.len() {
-            return (offset, offset);
-        }
-
-        // Find word boundaries
-        let mut start = offset;
-        while start > 0
-            && (chars[start - 1].is_alphanumeric()
-                || chars[start - 1] == '_'
-                || chars[start - 1] == '$'
-                || chars[start - 1] == '@'
-                || chars[start - 1] == '%')
-        {
-            start -= 1;
-        }
-
-        let mut end = offset;
-        while end < chars.len() && (chars[end].is_alphanumeric() || chars[end] == '_') {
-            end += 1;
-        }
-
-        (start, end)
+        Ok(Some(document_not_found_error()))
     }
 
     // === BEGIN_TEST_ONLY_POSITION_HELPERS ===
@@ -1567,304 +553,9 @@ impl LspServer {
         doc.line_starts.offset_to_position_rope(&doc.rope, offset)
     }
 
-    /// Handle textDocument/codeLens request
-    fn handle_code_lens(&self, params: Option<Value>) -> Result<Option<Value>, JsonRpcError> {
-        // Gate unadvertised feature
-        if !self.advertised_features.lock().unwrap().code_lens {
-            return Err(crate::lsp_errors::method_not_advertised());
-        }
-
-        if let Some(params) = params {
-            let uri = params["textDocument"]["uri"].as_str().unwrap_or("");
-
-            let documents = self.documents.lock().unwrap();
-            if let Some(doc) = self.get_document(&documents, uri) {
-                if let Some(ref ast) = doc.ast {
-                    let provider = CodeLensProvider::new(doc.text.clone());
-                    let mut lenses = provider.extract(ast);
-
-                    // Add shebang lens if applicable
-                    if let Some(shebang_lens) = get_shebang_lens(&doc.text) {
-                        lenses.insert(0, shebang_lens);
-                    }
-
-                    return Ok(Some(json!(lenses)));
-                } else {
-                    // Text-based fallback when AST is not available
-                    let text_lenses = self.extract_text_based_code_lenses(&doc.text, uri);
-                    return Ok(Some(json!(text_lenses)));
-                }
-            }
-        }
-
-        Ok(Some(json!([])))
-    }
-
-    /// Handle codeLens/resolve request
-    fn handle_code_lens_resolve(
-        &self,
-        params: Option<Value>,
-    ) -> Result<Option<Value>, JsonRpcError> {
-        if let Some(params) = params {
-            // Parse the code lens
-            if let Ok(lens) =
-                serde_json::from_value::<crate::code_lens_provider::CodeLens>(params.clone())
-            {
-                // Extract the symbol name and kind from the lens data
-                let symbol_name = lens
-                    .data
-                    .as_ref()
-                    .and_then(|d| d.get("name"))
-                    .and_then(|n| n.as_str())
-                    .unwrap_or("");
-
-                let symbol_kind = lens
-                    .data
-                    .as_ref()
-                    .and_then(|d| d.get("kind"))
-                    .and_then(|k| k.as_str())
-                    .unwrap_or("unknown");
-
-                // Get the document URI from the range (we need to track this better in the future)
-                // For now, count references across all documents
-                let mut total_references = 0;
-
-                let documents = self.documents.lock().unwrap();
-                for (_uri, doc) in documents.iter() {
-                    if let Some(ref ast) = doc.ast {
-                        total_references += self.count_references(ast, symbol_name, symbol_kind);
-                    } else {
-                        // Text-based fallback when AST is not available
-                        total_references +=
-                            self.count_references_text_based(&doc.text, symbol_name, symbol_kind);
-                    }
-                }
-
-                let resolved = resolve_code_lens(lens, total_references);
-                return Ok(Some(json!(resolved)));
-            }
-        }
-
-        Err(JsonRpcError { code: -32602, message: "Invalid parameters".to_string(), data: None })
-    }
-
-    /// Handle textDocument/inlineCompletion request
-    fn handle_inline_completion(
-        &self,
-        params: Option<Value>,
-    ) -> Result<Option<Value>, JsonRpcError> {
-        use crate::inline_completions::InlineCompletionProvider;
-
-        if let Some(params) = params {
-            let uri = params["textDocument"]["uri"].as_str().unwrap_or("");
-            let position = &params["position"];
-            let line = position["line"].as_u64().unwrap_or(0) as u32;
-            let character = position["character"].as_u64().unwrap_or(0) as u32;
-
-            let documents = self.documents.lock().unwrap();
-            if let Some(doc) = self.get_document(&documents, uri) {
-                let provider = InlineCompletionProvider::new();
-                let completions = provider.get_inline_completions(&doc.text, line, character);
-                return Ok(Some(serde_json::to_value(completions).unwrap_or(Value::Null)));
-            }
-        }
-
-        Ok(Some(json!({
-            "items": []
-        })))
-    }
-
-    /// Handle textDocument/inlineValue request
-    fn handle_inline_value(&self, params: Option<Value>) -> Result<Option<Value>, JsonRpcError> {
-        if let Some(params) = params {
-            let uri = params["textDocument"]["uri"].as_str().unwrap_or("");
-            let range = &params["range"];
-            let _context = &params["context"]; // Debug context (stopped at breakpoint, etc)
-
-            let documents = self.documents.lock().unwrap();
-            if let Some(doc) = self.get_document(&documents, uri) {
-                // Extract visible scalar variables in the range
-                let start_line = range["start"]["line"].as_u64().unwrap_or(0) as u32;
-                let end_line = range["end"]["line"].as_u64().unwrap_or(0) as u32;
-
-                let mut inline_values = Vec::new();
-
-                // Simple implementation: find scalar variables in the visible range
-                let lines: Vec<&str> = doc.text.lines().collect();
-                // Move regex construction outside loop
-                let re = regex::Regex::new(r"\$([a-zA-Z_][a-zA-Z0-9_]*)").unwrap();
-
-                for line_num in start_line..=end_line.min((lines.len() - 1) as u32) {
-                    let line_text = lines[line_num as usize];
-
-                    // Find scalar variables using regex
-                    for cap in re.captures_iter(line_text) {
-                        if let Some(m) = cap.get(0) {
-                            let var_text = m.as_str();
-                            let col = m.start();
-
-                            // Create inline value text hint (showing the variable name as placeholder)
-                            inline_values.push(json!({
-                                "range": {
-                                    "start": { "line": line_num, "character": col as u32 },
-                                    "end": { "line": line_num, "character": (col + var_text.len()) as u32 }
-                                },
-                                "text": format!("{} = ?", var_text)
-                            }));
-                        }
-                    }
-                }
-
-                return Ok(Some(json!(inline_values)));
-            }
-        }
-
-        Ok(Some(json!([])))
-    }
-
-    /// Handle textDocument/moniker request
-    fn handle_moniker(&self, params: Option<Value>) -> Result<Option<Value>, JsonRpcError> {
-        if let Some(params) = params {
-            let uri = params["textDocument"]["uri"].as_str().unwrap_or("");
-            let position = &params["position"];
-            let line = position["line"].as_u64().unwrap_or(0) as u32;
-            let character = position["character"].as_u64().unwrap_or(0) as u32;
-
-            let documents = self.documents.lock().unwrap();
-            if let Some(doc) = self.get_document(&documents, uri) {
-                if let Some(ref ast) = doc.ast {
-                    let offset = self.pos16_to_offset(doc, line, character);
-
-                    // Find the symbol at the cursor position
-                    let current_pkg = crate::declaration::current_package_at(ast, offset);
-                    if let Some(key) =
-                        crate::declaration::symbol_at_cursor(ast, offset, current_pkg)
-                    {
-                        // Generate a stable moniker for the symbol
-                        let identifier = format!("{}::{}", key.pkg, key.name).replace("::", ".");
-                        let moniker = json!({
-                            "scheme": "perl",
-                            "identifier": identifier,
-                            "unique": "project",
-                            "kind": "export"
-                        });
-
-                        return Ok(Some(json!([moniker])));
-                    }
-                }
-            }
-        }
-
-        Ok(Some(json!([])))
-    }
-
-    /// Handle textDocument/documentColor request
-    fn handle_document_color(&self, _params: Option<Value>) -> Result<Option<Value>, JsonRpcError> {
-        Err(JsonRpcError { code: -32601, message: "Method not found".into(), data: None })
-    }
-
-    /// Handle textDocument/colorPresentation request
-    fn handle_color_presentation(
-        &self,
-        _params: Option<Value>,
-    ) -> Result<Option<Value>, JsonRpcError> {
-        Err(JsonRpcError { code: -32601, message: "Method not found".into(), data: None })
-    }
-
-    /// Handle textDocument/linkedEditingRange request
-    fn handle_linked_editing_range(
-        &self,
-        params: Option<Value>,
-    ) -> Result<Option<Value>, JsonRpcError> {
-        // Gate unadvertised feature
-        if !self.advertised_features.lock().unwrap().linked_editing {
-            return Err(crate::lsp_errors::method_not_advertised());
-        }
-
-        if let Some(params) = params {
-            let uri = params["textDocument"]["uri"].as_str().unwrap_or("");
-            let position = &params["position"];
-            let line = position["line"].as_u64().unwrap_or(0) as u32;
-            let character = position["character"].as_u64().unwrap_or(0) as u32;
-
-            let documents = self.documents.lock().unwrap();
-            if let Some(doc) = self.get_document(&documents, uri) {
-                let result =
-                    crate::linked_editing::handle_linked_editing(&doc.text, line, character);
-                return Ok(Some(serde_json::to_value(result).unwrap_or(Value::Null)));
-            }
-        }
-
-        Ok(Some(Value::Null))
-    }
-
-    /// Count references to a symbol using text-based search
-    fn count_references_text_based(
-        &self,
-        text: &str,
-        symbol_name: &str,
-        symbol_kind: &str,
-    ) -> usize {
-        let mut count = 0;
-
-        match symbol_kind {
-            "package" => {
-                // Count package usage (use statements, new() calls, etc.)
-                use regex::Regex;
-
-                // Count "use PackageName" statements
-                if let Ok(use_regex) =
-                    Regex::new(&format!(r"\buse\s+{}\b", regex::escape(symbol_name)))
-                {
-                    count += use_regex.find_iter(text).count();
-                }
-
-                // Count "PackageName->new()" or "PackageName->method()" calls
-                if let Ok(call_regex) = Regex::new(&format!(r"\b{}->", regex::escape(symbol_name)))
-                {
-                    count += call_regex.find_iter(text).count();
-                }
-
-                // Count "bless ... PackageName" statements
-                if let Ok(bless_regex) =
-                    Regex::new(&format!(r"bless\s+.*?,\s*{}", regex::escape(symbol_name)))
-                {
-                    count += bless_regex.find_iter(text).count();
-                }
-            }
-            "subroutine" => {
-                // Count function calls
-                use regex::Regex;
-
-                // Count "function_name(" calls
-                if let Ok(call_regex) =
-                    Regex::new(&format!(r"\b{}\s*\(", regex::escape(symbol_name)))
-                {
-                    count += call_regex.find_iter(text).count();
-                }
-
-                // Count "&function_name" references
-                if let Ok(ref_regex) = Regex::new(&format!(r"&{}\b", regex::escape(symbol_name))) {
-                    count += ref_regex.find_iter(text).count();
-                }
-            }
-            _ => {
-                // Generic symbol name search for unknown kinds
-                use regex::Regex;
-                if let Ok(generic_regex) =
-                    Regex::new(&format!(r"\b{}\b", regex::escape(symbol_name)))
-                {
-                    count += generic_regex.find_iter(text).count();
-                    // Don't count the definition itself if it exists
-                    if count > 0 {
-                        count = count.saturating_sub(1);
-                    }
-                }
-            }
-        }
-
-        count
-    }
+    // Note: handle_code_lens, handle_code_lens_resolve, handle_inline_completion,
+    // handle_inline_value, handle_moniker, handle_document_color, handle_color_presentation,
+    // handle_linked_editing_range, count_references_text_based are implemented in language/misc.rs
 
     /// Extract code lenses from text when AST parsing fails
     fn extract_text_based_code_lenses(
@@ -2402,208 +1093,11 @@ impl LspServer {
         count
     }
 
-    /// Handle semantic tokens full request
-    fn handle_semantic_tokens_full(
-        &self,
-        params: Option<Value>,
-    ) -> Result<Option<Value>, JsonRpcError> {
-        if let Some(params) = params {
-            let uri = params["textDocument"]["uri"].as_str().unwrap_or("");
+    // Note: handle_semantic_tokens_full, handle_semantic_tokens_range are implemented
+    // in language/semantic_tokens.rs
 
-            eprintln!("Getting semantic tokens for: {}", uri);
-
-            let documents = self.documents.lock().unwrap();
-            if let Some(doc) = self.get_document(&documents, uri) {
-                if let Some(ref ast) = doc.ast {
-                    let provider = SemanticTokensProvider::new(doc.text.clone());
-                    let tokens = provider.extract(ast);
-                    let encoded = encode_semantic_tokens(&tokens);
-
-                    eprintln!("Found {} semantic tokens", tokens.len());
-
-                    return Ok(Some(json!({
-                        "data": encoded
-                    })));
-                }
-            }
-        }
-
-        Ok(Some(json!({
-            "data": []
-        })))
-    }
-
-    /// Handle semantic tokens range request
-    fn handle_semantic_tokens_range(
-        &self,
-        params: Option<Value>,
-    ) -> Result<Option<Value>, JsonRpcError> {
-        if let Some(params) = params {
-            let uri = params["textDocument"]["uri"].as_str().unwrap_or("");
-            let range = &params["range"];
-            let start_line = range["start"]["line"].as_u64().unwrap_or(0) as u32;
-            let end_line = range["end"]["line"].as_u64().unwrap_or(u32::MAX as u64) as u32;
-
-            eprintln!(
-                "Getting semantic tokens for range: {} (lines {}-{})",
-                uri, start_line, end_line
-            );
-
-            let documents = self.documents.lock().unwrap();
-            if let Some(doc) = self.get_document(&documents, uri) {
-                if let Some(ref ast) = doc.ast {
-                    let provider = SemanticTokensProvider::new(doc.text.clone());
-                    let all_tokens = provider.extract(ast);
-
-                    // Filter tokens to the requested range
-                    let range_tokens: Vec<_> = all_tokens
-                        .into_iter()
-                        .filter(|token| token.line >= start_line && token.line <= end_line)
-                        .collect();
-
-                    let encoded = encode_semantic_tokens(&range_tokens);
-
-                    eprintln!("Found {} semantic tokens in range", range_tokens.len());
-
-                    return Ok(Some(json!({
-                        "data": encoded
-                    })));
-                }
-            }
-        }
-
-        Ok(Some(json!({
-            "data": []
-        })))
-    }
-
-    /// Handle prepare call hierarchy request
-    fn handle_prepare_call_hierarchy(
-        &self,
-        params: Option<Value>,
-    ) -> Result<Option<Value>, JsonRpcError> {
-        // Gate unadvertised feature
-        if !self.advertised_features.lock().unwrap().call_hierarchy {
-            return Err(crate::lsp_errors::method_not_advertised());
-        }
-
-        if let Some(params) = params {
-            let uri = params["textDocument"]["uri"].as_str().unwrap_or("");
-            let position = &params["position"];
-            let line = position["line"].as_u64().unwrap_or(0) as u32;
-            let character = position["character"].as_u64().unwrap_or(0) as u32;
-
-            eprintln!("Preparing call hierarchy at: {} ({}:{})", uri, line, character);
-
-            let documents = self.documents.lock().unwrap();
-            if let Some(doc) = self.get_document(&documents, uri) {
-                if let Some(ref ast) = doc.ast {
-                    let provider = CallHierarchyProvider::new(doc.text.clone(), uri.to_string());
-                    if let Some(items) = provider.prepare(ast, line, character) {
-                        let json_items: Vec<_> = items.iter().map(|item| item.to_json()).collect();
-                        return Ok(Some(json!(json_items)));
-                    }
-                }
-            }
-        }
-
-        Ok(Some(json!(null)))
-    }
-
-    /// Handle incoming calls request
-    fn handle_incoming_calls(&self, params: Option<Value>) -> Result<Option<Value>, JsonRpcError> {
-        if let Some(params) = params {
-            let item = &params["item"];
-            let uri = item["uri"].as_str().unwrap_or("");
-
-            eprintln!("Getting incoming calls for: {}", item["name"].as_str().unwrap_or(""));
-
-            let documents = self.documents.lock().unwrap();
-            if let Some(doc) = self.get_document(&documents, uri) {
-                if let Some(ref ast) = doc.ast {
-                    // Reconstruct the CallHierarchyItem from JSON
-                    let ch_item = self.json_to_call_hierarchy_item(item)?;
-
-                    let provider = CallHierarchyProvider::new(doc.text.clone(), uri.to_string());
-                    let calls = provider.incoming_calls(ast, &ch_item);
-
-                    let json_calls: Vec<_> = calls.iter().map(|call| call.to_json()).collect();
-                    return Ok(Some(json!(json_calls)));
-                }
-            }
-        }
-
-        Ok(Some(json!([])))
-    }
-
-    /// Handle outgoing calls request
-    fn handle_outgoing_calls(&self, params: Option<Value>) -> Result<Option<Value>, JsonRpcError> {
-        if let Some(params) = params {
-            let item = &params["item"];
-            let uri = item["uri"].as_str().unwrap_or("");
-
-            eprintln!("Getting outgoing calls for: {}", item["name"].as_str().unwrap_or(""));
-
-            let documents = self.documents.lock().unwrap();
-            if let Some(doc) = self.get_document(&documents, uri) {
-                if let Some(ref ast) = doc.ast {
-                    // Reconstruct the CallHierarchyItem from JSON
-                    let ch_item = self.json_to_call_hierarchy_item(item)?;
-
-                    let provider = CallHierarchyProvider::new(doc.text.clone(), uri.to_string());
-                    let calls = provider.outgoing_calls(ast, &ch_item);
-
-                    let json_calls: Vec<_> = calls.iter().map(|call| call.to_json()).collect();
-                    return Ok(Some(json!(json_calls)));
-                }
-            }
-        }
-
-        Ok(Some(json!([])))
-    }
-
-    /// Convert JSON to CallHierarchyItem
-    fn json_to_call_hierarchy_item(
-        &self,
-        json: &Value,
-    ) -> Result<crate::call_hierarchy_provider::CallHierarchyItem, JsonRpcError> {
-        use crate::call_hierarchy_provider::{CallHierarchyItem, Position, Range};
-
-        let name = json["name"].as_str().unwrap_or("").to_string();
-        let kind = match json["kind"].as_u64().unwrap_or(12) {
-            6 => "method",
-            _ => "function",
-        }
-        .to_string();
-        let uri = json["uri"].as_str().unwrap_or("").to_string();
-
-        let range = Range {
-            start: Position {
-                line: json["range"]["start"]["line"].as_u64().unwrap_or(0) as u32,
-                character: json["range"]["start"]["character"].as_u64().unwrap_or(0) as u32,
-            },
-            end: Position {
-                line: json["range"]["end"]["line"].as_u64().unwrap_or(0) as u32,
-                character: json["range"]["end"]["character"].as_u64().unwrap_or(0) as u32,
-            },
-        };
-
-        let selection_range = Range {
-            start: Position {
-                line: json["selectionRange"]["start"]["line"].as_u64().unwrap_or(0) as u32,
-                character: json["selectionRange"]["start"]["character"].as_u64().unwrap_or(0)
-                    as u32,
-            },
-            end: Position {
-                line: json["selectionRange"]["end"]["line"].as_u64().unwrap_or(0) as u32,
-                character: json["selectionRange"]["end"]["character"].as_u64().unwrap_or(0) as u32,
-            },
-        };
-
-        let detail = json["detail"].as_str().map(|s| s.to_string());
-
-        Ok(CallHierarchyItem { name, kind, uri, range, selection_range, detail })
-    }
+    // Note: handle_prepare_call_hierarchy, handle_incoming_calls, handle_outgoing_calls,
+    // json_to_call_hierarchy_item are implemented in language/hierarchy.rs
 
     /// Find matching closing parenthesis
     fn find_matching_paren(&self, s: &str, open_at: usize) -> Option<usize> {
@@ -2876,669 +1370,8 @@ impl LspServer {
         (s, e, &text[s.min(text.len())..e.min(text.len())])
     }
 
-    /// Handle inlay hint request
-    fn handle_inlay_hint(&self, params: Option<Value>) -> Result<Option<Value>, JsonRpcError> {
-        if let Some(params) = params {
-            let uri = params["textDocument"]["uri"].as_str().unwrap_or("");
-            let range = &params["range"];
-
-            eprintln!("Getting inlay hints for: {}", uri);
-
-            let documents = self.documents.lock().unwrap();
-            if let Some(doc) = self.get_document(&documents, uri) {
-                let mut hints = Vec::new();
-
-                if let Some(ref ast) = doc.ast {
-                    // Configure inlay hints based on server settings
-                    let server_config = self.config.lock().unwrap();
-                    let config = InlayHintConfig {
-                        parameter_hints: server_config.inlay_hints_parameter_hints,
-                        type_hints: server_config.inlay_hints_type_hints,
-                        chained_hints: server_config.inlay_hints_chained_hints,
-                        max_length: server_config.inlay_hints_max_length,
-                    };
-
-                    let provider = InlayHintsProvider::with_config(doc.text.clone(), config);
-
-                    // Extract range if provided
-                    let lsp_range = if params.get("range").is_some() {
-                        let start_line = range["start"]["line"].as_u64().unwrap_or(0) as u32;
-                        let start_char = range["start"]["character"].as_u64().unwrap_or(0) as u32;
-                        let end_line =
-                            range["end"]["line"].as_u64().unwrap_or(u32::MAX as u64) as u32;
-                        let end_char =
-                            range["end"]["character"].as_u64().unwrap_or(u32::MAX as u64) as u32;
-                        Some(crate::positions::Range::new(
-                            start_line, start_char, end_line, end_char,
-                        ))
-                    } else {
-                        None
-                    };
-
-                    // Use the extract_range method if range is provided
-                    let ast_hints = if let Some(r) = lsp_range {
-                        provider.extract_range(ast, r)
-                    } else {
-                        provider.extract(ast)
-                    };
-
-                    for hint in ast_hints {
-                        hints.push(hint.to_json());
-                    }
-                }
-
-                // Named-arg fallback: foo(bar => $x, 'baz' => 2)
-                // Respect client-supplied range if present
-                let has_range = params.get("range").and_then(|r| r.as_object()).is_some();
-                let (_win_s, _win_e, window) = if has_range {
-                    let s_line = range["start"]["line"].as_u64().unwrap_or(0) as u32;
-                    let s_ch = range["start"]["character"].as_u64().unwrap_or(0) as u32;
-                    let e_line = range["end"]["line"].as_u64().unwrap_or(0) as u32;
-                    let e_ch = range["end"]["character"].as_u64().unwrap_or(0) as u32;
-                    self.slice_in_range(&doc.text, (s_line, s_ch), (e_line, e_ch))
-                } else {
-                    (0, doc.text.len(), doc.text.as_str())
-                };
-
-                // Add named argument hints for => pairs
-                use crate::builtin_signatures_phf::{BUILTIN_SIGS, get_param_names};
-                use regex::Regex;
-                use std::collections::HashSet;
-                lazy_static::lazy_static! {
-                    static ref RE_CALL: Regex = Regex::new(r"(?m)([A-Za-z_]\w*)\s*\(").unwrap();
-                    static ref RE_PAIR: Regex = Regex::new(
-                        r#"(?x)
-                        (?P<key>
-                            [A-Za-z_]\w*                # bareword
-                            | '([^'\\]|\\.)*'           # single-quoted
-                            | "([^"\\]|\\.)*"           # double-quoted
-                        )
-                        \s*=>\s*
-                        "#
-                    ).unwrap();
-                }
-
-                for m in RE_PAIR.find_iter(window) {
-                    if let Some(caps) = RE_PAIR.captures(m.as_str()) {
-                        let mut key =
-                            caps.name("key").map(|k| k.as_str()).unwrap_or("").to_string();
-
-                        // Strip quotes if present
-                        if (key.starts_with('"') && key.ends_with('"'))
-                            || (key.starts_with('\'') && key.ends_with('\''))
-                        {
-                            key = key[1..key.len().saturating_sub(1)].to_string();
-                        }
-
-                        // Calculate position for the hint (at the value start)
-                        let val_offset = m.end();
-                        let (l, c) = doc.line_starts.offset_to_position_rope(&doc.rope, val_offset);
-
-                        // Only add hint if within requested range
-                        let start_line = range["start"]["line"].as_u64().unwrap_or(0) as u32;
-                        let start_char = range["start"]["character"].as_u64().unwrap_or(0) as u32;
-                        let end_line =
-                            range["end"]["line"].as_u64().unwrap_or(u32::MAX as u64) as u32;
-                        let end_char =
-                            range["end"]["character"].as_u64().unwrap_or(u32::MAX as u64) as u32;
-
-                        let pos = crate::positions::Position::new(l, c);
-                        let range = crate::positions::Range::new(
-                            start_line, start_char, end_line, end_char,
-                        );
-                        if crate::positions::pos_in_range(pos, range) {
-                            hints.push(json!({
-                                "position": { "line": l, "character": c },
-                                "label": format!("{}:", key),
-                                "kind": 2, // Parameter hint
-                            }));
-                        }
-                    }
-                }
-
-                // ---- Parameter hints for common built-ins with comma-separated args ----
-                // Dedup across AST/named-arg hints
-                let mut seen: HashSet<(u32, u32, String)> = HashSet::new();
-                for h in &hints {
-                    if let (Some(l), Some(c), Some(lbl)) = (
-                        h.get("position")
-                            .and_then(|p| p.get("line"))
-                            .and_then(|v| v.as_u64())
-                            .map(|v| v as u32),
-                        h.get("position")
-                            .and_then(|p| p.get("character"))
-                            .and_then(|v| v.as_u64())
-                            .map(|v| v as u32),
-                        h.get("label").and_then(|v| v.as_str()).map(|s| s.to_string()),
-                    ) {
-                        seen.insert((l, c, lbl));
-                    }
-                }
-
-                // Range window (reuse earlier computed range/window start)
-                let has_range_2 = params.get("range").and_then(|r| r.as_object()).is_some();
-                let (win_s, _win_e2, window2) = if has_range_2 {
-                    let s_line = range["start"]["line"].as_u64().unwrap_or(0) as u32;
-                    let s_ch = range["start"]["character"].as_u64().unwrap_or(0) as u32;
-                    let e_line = range["end"]["line"].as_u64().unwrap_or(0) as u32;
-                    let e_ch = range["end"]["character"].as_u64().unwrap_or(0) as u32;
-                    self.slice_in_range(&doc.text, (s_line, s_ch), (e_line, e_ch))
-                } else {
-                    (0, doc.text.len(), doc.text.as_str())
-                };
-                let _start_line = range["start"]["line"].as_u64().unwrap_or(0) as u32;
-                let _end_line = range["end"]["line"].as_u64().unwrap_or(u32::MAX as u64) as u32;
-
-                for m in RE_CALL.find_iter(window2) {
-                    // function name
-                    let caps = RE_CALL.captures(&window2[m.start()..m.end()]).unwrap();
-                    let name = caps.get(1).unwrap().as_str();
-                    let sig = get_param_names(name);
-                    if sig.is_empty() {
-                        continue;
-                    }
-                    // find the matching ')' in the **whole window** after '('
-                    // compute open and close offsets relative to window
-                    let open_global = win_s + m.end() - 1; // m ends right after '(' due to regex; step back to '('
-                    let open_in_window = m.end() - 1;
-                    let close_in_window = match self.find_matching_paren(window2, open_in_window) {
-                        Some(p) => p,
-                        None => continue,
-                    };
-                    let body = &window2[open_in_window + 1..close_in_window];
-                    let arg_starts = self.arg_starts_in_call_body(body);
-                    for (i, rel) in arg_starts.iter().enumerate() {
-                        if i >= sig.len() {
-                            break;
-                        }
-                        let local_anchor = self.smart_arg_anchor(body, *rel);
-                        let global_off = open_global + 1 + local_anchor;
-                        let (l, c) = doc.line_starts.offset_to_position_rope(&doc.rope, global_off);
-                        // Get range bounds for position checking
-                        let start_line = range["start"]["line"].as_u64().unwrap_or(0) as u32;
-                        let start_char = range["start"]["character"].as_u64().unwrap_or(0) as u32;
-                        let end_line =
-                            range["end"]["line"].as_u64().unwrap_or(u32::MAX as u64) as u32;
-                        let end_char =
-                            range["end"]["character"].as_u64().unwrap_or(u32::MAX as u64) as u32;
-                        let pos = crate::positions::Position::new(l, c);
-                        let hint_range = crate::positions::Range::new(
-                            start_line, start_char, end_line, end_char,
-                        );
-                        if !crate::positions::pos_in_range(pos, hint_range) {
-                            continue;
-                        }
-                        let label = format!("{}:", sig[i]);
-                        if seen.insert((l, c, label.clone())) {
-                            hints.push(json!({
-                                "position": { "line": l, "character": c },
-                                "label": label,
-                                "kind": 2, // Parameter
-                            }));
-                        }
-                    }
-                }
-
-                // Non-parenthesized built-ins: `push @arr, ...;`, `open FH, ...;`, `split /re/, s;`
-                for (name, sig) in BUILTIN_SIGS.entries() {
-                    let re = Regex::new(&format!(r#"(?m)\b{}\b(?!\s*\()"#, regex::escape(name)))
-                        .unwrap();
-                    for m in re.find_iter(window2) {
-                        // body = from end of name to end-of-statement `;` (window-limited)
-                        let body_start = m.end();
-                        let stmt_end_in_win = self.slice_until_stmt_end(window2, body_start);
-                        let body = &window2[body_start..stmt_end_in_win];
-                        let arg_starts = self.arg_starts_top_level(body);
-                        for (i, rel) in arg_starts.iter().enumerate() {
-                            if i >= sig.len() {
-                                break;
-                            }
-                            // Re-anchor on the variable/filehandle token (skip `my`/`our`, whitespace)
-                            let local_anchor = self.smart_arg_anchor(body, *rel);
-                            let global_off = win_s + body_start + local_anchor;
-                            let (l, c) =
-                                doc.line_starts.offset_to_position_rope(&doc.rope, global_off);
-                            // Get range bounds for position checking
-                            let start_line = range["start"]["line"].as_u64().unwrap_or(0) as u32;
-                            let start_char =
-                                range["start"]["character"].as_u64().unwrap_or(0) as u32;
-                            let end_line =
-                                range["end"]["line"].as_u64().unwrap_or(u32::MAX as u64) as u32;
-                            let end_char =
-                                range["end"]["character"].as_u64().unwrap_or(u32::MAX as u64)
-                                    as u32;
-                            let pos = crate::positions::Position::new(l, c);
-                            let hint_range = crate::positions::Range::new(
-                                start_line, start_char, end_line, end_char,
-                            );
-                            if !crate::positions::pos_in_range(pos, hint_range) {
-                                continue;
-                            }
-                            let label = format!("{}:", sig[i]);
-                            if seen.insert((l, c, label.clone())) {
-                                hints.push(json!({
-                                    "position": { "line": l, "character": c },
-                                    "label": label,
-                                    "kind": 2, // Parameter
-                                }));
-                            }
-                        }
-                    }
-                }
-
-                // ---- Type hints (simple heuristics) ----
-                // Patterns:
-                //  my %hash;              -> "hash"
-                //  my @arr;               -> "array"
-                //  my $ref = {};          -> "HashRef"
-                //  my $ref = [];          -> "ArrayRef"
-                //  my $s   = "text";      -> "Str"
-                let type_patterns: &[(&str, &str, u8)] = &[
-                    // allow optional initializer: `my %h;` or `my %h = ();`
-                    (r"(?m)my\s+%([A-Za-z_]\w*)\b(?:\s*=\s*\(\s*\))?", "hash", 1),
-                    (r"(?m)my\s+@([A-Za-z_]\w*)\b(?:\s*=\s*\(\s*\))?", "array", 1),
-                    (r"(?m)my\s+\$([A-Za-z_]\w*)\s*=\s*\{", "HashRef", 1),
-                    (r"(?m)my\s+\$([A-Za-z_]\w*)\s*=\s*\[", "ArrayRef", 1),
-                    (r#"(?m)my\s+\$([A-Za-z_]\w*)\s*=\s*"(?:[^"\\]|\\.)*""#, "Str", 1),
-                ];
-                for (pat, label_txt, kind_code) in type_patterns {
-                    let re = Regex::new(pat).unwrap();
-                    for m in re.find_iter(window2) {
-                        // Position hint at the variable's start (first capture)
-                        if let Some(caps) = re.captures(&window2[m.start()..m.end()]) {
-                            if let Some(var) = caps.get(1) {
-                                let var_global = win_s + m.start() + var.start();
-                                let (l, c) =
-                                    doc.line_starts.offset_to_position_rope(&doc.rope, var_global);
-                                // Get range bounds for position checking
-                                let start_line =
-                                    range["start"]["line"].as_u64().unwrap_or(0) as u32;
-                                let start_char =
-                                    range["start"]["character"].as_u64().unwrap_or(0) as u32;
-                                let end_line =
-                                    range["end"]["line"].as_u64().unwrap_or(u32::MAX as u64) as u32;
-                                let end_char =
-                                    range["end"]["character"].as_u64().unwrap_or(u32::MAX as u64)
-                                        as u32;
-                                let pos = crate::positions::Position::new(l, c);
-                                let hint_range = crate::positions::Range::new(
-                                    start_line, start_char, end_line, end_char,
-                                );
-                                if !crate::positions::pos_in_range(pos, hint_range) {
-                                    continue;
-                                }
-                                let label = label_txt.to_string();
-                                if seen.insert((l, c, label.clone())) {
-                                    hints.push(json!({
-                                        "position": { "line": l, "character": c },
-                                        "label": label,
-                                        "kind": *kind_code, // 1 = Type
-                                    }));
-                                }
-                            }
-                        }
-                    }
-                }
-
-                return Ok(Some(json!(hints)));
-            }
-        }
-
-        Ok(Some(json!([])))
-    }
-
-    /// Handle test discovery request
-    fn handle_test_discovery(&self, params: Option<Value>) -> Result<Option<Value>, JsonRpcError> {
-        if let Some(params) = params {
-            let uri = params["textDocument"]["uri"].as_str().unwrap_or("");
-
-            eprintln!("Discovering tests for: {}", uri);
-
-            let documents = self.documents.lock().unwrap();
-            if let Some(doc) = self.get_document(&documents, uri) {
-                if let Some(ref ast) = doc.ast {
-                    let runner = TestRunner::new(doc.text.clone(), uri.to_string());
-                    let tests = runner.discover_tests(ast);
-
-                    // Convert test items to JSON
-                    let test_items: Vec<Value> = tests
-                        .into_iter()
-                        .map(|test| {
-                            json!({
-                                "id": test.id,
-                                "label": test.label,
-                                "uri": test.uri,
-                                "range": {
-                                    "start": {
-                                        "line": test.range.start_line,
-                                        "character": test.range.start_character
-                                    },
-                                    "end": {
-                                        "line": test.range.end_line,
-                                        "character": test.range.end_character
-                                    }
-                                },
-                                "kind": match test.kind {
-                                    TestKind::File => "file",
-                                    TestKind::Suite => "suite",
-                                    TestKind::Test => "test"
-                                },
-                                "children": test.children.into_iter()
-                                    .map(|child| json!({
-                                        "id": child.id,
-                                        "label": child.label,
-                                        "uri": child.uri,
-                                        "range": {
-                                            "start": {
-                                                "line": child.range.start_line,
-                                                "character": child.range.start_character
-                                            },
-                                            "end": {
-                                                "line": child.range.end_line,
-                                                "character": child.range.end_character
-                                            }
-                                        },
-                                        "kind": match child.kind {
-                                            TestKind::File => "file",
-                                            TestKind::Suite => "suite",
-                                            TestKind::Test => "test"
-                                        },
-                                        "children": []
-                                    }))
-                                    .collect::<Vec<_>>()
-                            })
-                        })
-                        .collect();
-
-                    eprintln!("Found {} test items", test_items.len());
-
-                    return Ok(Some(json!(test_items)));
-                }
-            }
-        }
-
-        Ok(Some(json!([])))
-    }
-
-    /// Handle execute command request
-    fn handle_execute_command(&self, params: Option<Value>) -> Result<Option<Value>, JsonRpcError> {
-        use crate::execute_command::ExecuteCommandProvider;
-
-        if let Some(params) = params {
-            let command = params["command"].as_str().unwrap_or("");
-
-            // LSP 3.17 compliance: arguments field is required even if empty
-            if !params.as_object().unwrap_or(&serde_json::Map::new()).contains_key("arguments") {
-                return Err(JsonRpcError {
-                    code: -32602, // InvalidParams
-                    message: "Missing required 'arguments' field in executeCommand request"
-                        .to_string(),
-                    data: Some(json!({
-                        "command": command,
-                        "errorType": "executeCommand",
-                        "originalError": "Missing 'arguments' field"
-                    })),
-                });
-            }
-
-            let arguments = params["arguments"].as_array().cloned().unwrap_or_default();
-
-            eprintln!("Executing command: {}", command);
-
-            // Use the new execute command provider for new commands
-            let provider = ExecuteCommandProvider::new();
-
-            match command {
-                // Keep existing test commands for backward compatibility
-                "perl.runTest" => {
-                    if let Some(test_id) = arguments.first().and_then(|v| v.as_str()) {
-                        return self.run_test(test_id);
-                    }
-                }
-                "perl.runTestFile" => {
-                    if let Some(file_uri) = arguments.first().and_then(|v| v.as_str()) {
-                        return self.run_test_file(file_uri);
-                    }
-                }
-                // New commands handled by ExecuteCommandProvider
-                "perl.runTests" | "perl.runFile" | "perl.runTestSub" | "perl.debugTests"
-                | "perl.runCritic" => {
-                    match provider.execute_command(command, arguments) {
-                        Ok(result) => return Ok(Some(result)),
-                        Err(e) => {
-                            // Return proper JSON-RPC error according to LSP 3.17 specification
-                            let error_code = if e.contains("Missing") || e.contains("argument") {
-                                -32602 // InvalidParams
-                            } else if e.contains("Unknown command") {
-                                -32601 // MethodNotFound
-                            } else if e.contains("Path traversal") || e.contains("security") {
-                                -32603 // InternalError (security)
-                            } else {
-                                -32603 // InternalError (general)
-                            };
-
-                            return Err(JsonRpcError {
-                                code: error_code,
-                                message: format!("Execute command failed: {}", e),
-                                data: Some(json!({
-                                    "command": command,
-                                    "errorType": "executeCommand",
-                                    "originalError": e
-                                })),
-                            });
-                        }
-                    }
-                }
-                // Debug commands (stub implementation for now)
-                "perl.debugFile" => {
-                    eprintln!("Debug command requested: {}", command);
-                    // Return a success status - actual DAP integration can be added later
-                    return Ok(Some(
-                        json!({"status": "started", "message": format!("Debug session {} initiated", command)}),
-                    ));
-                }
-                _ => {
-                    return Err(JsonRpcError {
-                        code: METHOD_NOT_FOUND,
-                        message: format!("Unknown command: {}", command),
-                        data: None,
-                    });
-                }
-            }
-        }
-
-        // Missing params entirely
-        Err(JsonRpcError {
-            code: -32602, // InvalidParams
-            message: "Missing parameters for executeCommand request".to_string(),
-            data: Some(json!({
-                "errorType": "executeCommand",
-                "originalError": "Missing params"
-            })),
-        })
-    }
-
-    /// Run a specific test
-    fn run_test(&self, test_id: &str) -> Result<Option<Value>, JsonRpcError> {
-        eprintln!("Running test: {}", test_id);
-
-        // Parse test ID to get URI and test name
-        let parts: Vec<&str> = test_id.split("::").collect();
-        if parts.len() < 2 {
-            return Ok(Some(json!({"status": "error", "message": "Invalid test ID"})));
-        }
-
-        let uri = parts[0];
-        let test_name = parts[1..].join("::");
-
-        let documents = self.documents.lock().unwrap();
-        if let Some(doc) = documents.get(uri) {
-            let runner = TestRunner::new(doc.text.clone(), uri.to_string());
-            let results = runner.run_test(&test_name);
-
-            // Convert results to JSON
-            let json_results: Vec<Value> = results
-                .into_iter()
-                .map(|result| {
-                    json!({
-                        "testId": result.test_id,
-                        "status": result.status.as_str(),
-                        "message": result.message,
-                        "duration": result.duration
-                    })
-                })
-                .collect();
-
-            return Ok(Some(json!({
-                "status": "success",
-                "results": json_results
-            })));
-        }
-
-        Ok(Some(document_not_found_error()))
-    }
-
-    /// Run all tests in a file
-    fn run_test_file(&self, uri: &str) -> Result<Option<Value>, JsonRpcError> {
-        eprintln!("Running test file: {}", uri);
-
-        let documents = self.documents.lock().unwrap();
-        if let Some(doc) = documents.get(uri) {
-            let runner = TestRunner::new(doc.text.clone(), uri.to_string());
-            let results = runner.run_test(uri);
-
-            // Convert results to JSON
-            let json_results: Vec<Value> = results
-                .into_iter()
-                .map(|result| {
-                    json!({
-                        "testId": result.test_id,
-                        "status": result.status.as_str(),
-                        "message": result.message,
-                        "duration": result.duration
-                    })
-                })
-                .collect();
-
-            return Ok(Some(json!({
-                "status": "success",
-                "results": json_results
-            })));
-        }
-
-        Ok(Some(document_not_found_error()))
-    }
-
-    /// Run Perl::Critic on a file
-    fn run_perl_critic(&self, uri: &str) -> Result<Option<Value>, JsonRpcError> {
-        use crate::perl_critic::{CriticAnalyzer, CriticConfig};
-        use std::path::Path;
-
-        eprintln!("Running Perl::Critic on: {}", uri);
-
-        // Get the document content
-        let documents = self.documents.lock().unwrap();
-        if let Some(doc) = documents.get(uri) {
-            // Try to get file path from URI
-            let file_path = if uri.starts_with("file://") {
-                uri.strip_prefix("file://").unwrap_or(uri)
-            } else {
-                uri
-            };
-
-            // First, try external perlcritic if available
-            let violations = if Path::new("/usr/bin/perlcritic").exists()
-                || Path::new("/usr/local/bin/perlcritic").exists()
-                || std::process::Command::new("which")
-                    .arg("perlcritic")
-                    .output()
-                    .map(|o| o.status.success())
-                    .unwrap_or(false)
-            {
-                // Use external perlcritic
-                let mut analyzer = CriticAnalyzer::new(CriticConfig::default());
-                match analyzer.analyze_file(Path::new(file_path)) {
-                    Ok(violations) => violations,
-                    Err(e) => {
-                        eprintln!("External perlcritic failed: {}, using built-in analyzer", e);
-                        // Fall back to built-in analyzer
-                        let builtin = BuiltInAnalyzer::new();
-                        let code_text = crate::util::code_slice(&doc.text);
-                        let mut parser = Parser::new(code_text);
-                        if let Ok(ast) = parser.parse() {
-                            builtin.analyze(&ast, &doc.text)
-                        } else {
-                            builtin.analyze(
-                                &Node::new(
-                                    NodeKind::Error { message: "Parse error".to_string() },
-                                    crate::ast::SourceLocation { start: 0, end: 0 },
-                                ),
-                                &doc.text,
-                            )
-                        }
-                    }
-                }
-            } else {
-                // Use built-in analyzer
-                eprintln!("Using built-in Perl::Critic analyzer");
-                let builtin = BuiltInAnalyzer::new();
-                let code_text = crate::util::code_slice(&doc.text);
-                let mut parser = Parser::new(code_text);
-                if let Ok(ast) = parser.parse() {
-                    builtin.analyze(&ast, &doc.text)
-                } else {
-                    builtin.analyze(
-                        &Node::new(
-                            NodeKind::Error { message: "Parse error".to_string() },
-                            crate::ast::SourceLocation { start: 0, end: 0 },
-                        ),
-                        &doc.text,
-                    )
-                }
-            };
-
-            // Convert violations to diagnostics and publish them
-            let diagnostics: Vec<Value> = violations.iter().map(|v| {
-                json!({
-                    "range": {
-                        "start": {"line": v.range.start.line, "character": v.range.start.column},
-                        "end": {"line": v.range.end.line, "character": v.range.end.column}
-                    },
-                    "severity": match v.severity {
-                        crate::perl_critic::Severity::Brutal |
-                        crate::perl_critic::Severity::Cruel => 1, // Error
-                        crate::perl_critic::Severity::Harsh => 2, // Warning
-                        _ => 3, // Information
-                    },
-                    "code": v.policy,
-                    "source": "Perl::Critic",
-                    "message": v.description
-                })
-            }).collect();
-
-            // Send publishDiagnostics notification
-            let _ = self.notify(
-                "textDocument/publishDiagnostics",
-                json!({
-                    "uri": uri,
-                    "diagnostics": diagnostics
-                }),
-            );
-
-            return Ok(Some(json!({
-                "status": "success",
-                "violationCount": violations.len(),
-                "violations": violations.iter().map(|v| json!({
-                    "policy": v.policy,
-                    "severity": format!("{:?}", v.severity),
-                    "message": v.description,
-                    "line": v.range.start.line + 1
-                })).collect::<Vec<_>>()
-            })));
-        }
-
-        Ok(Some(document_not_found_error()))
-    }
+    // Note: handle_inlay_hint, handle_test_discovery, handle_execute_command, run_test,
+    // run_test_file, run_perl_critic are implemented in language/misc.rs
 
     // Test-only public methods (enabled for unit tests or integration tests with expose_lsp_test_api)
     #[cfg(any(test, feature = "expose_lsp_test_api"))]
@@ -3570,29 +1403,7 @@ impl LspServer {
         self.handle_completion(params)
     }
 
-    /// Handle documentLink request
-    fn handle_document_link(&self, params: Option<Value>) -> Result<Option<Value>, JsonRpcError> {
-        if let Some(params) = params {
-            let uri = params["textDocument"]["uri"].as_str().unwrap_or("");
-
-            let documents = self.documents.lock().unwrap();
-            if let Some(doc) = self.get_document(&documents, uri) {
-                let uri_parsed = url::Url::parse(uri).map_err(|_| JsonRpcError {
-                    code: -32602,
-                    message: "Invalid URI".to_string(),
-                    data: None,
-                })?;
-                match crate::lsp_document_link::collect_document_links(&doc.text, &uri_parsed) {
-                    Ok(links) => Ok(Some(serde_json::to_value(links).unwrap_or(Value::Null))),
-                    Err(_) => Ok(Some(Value::Null)),
-                }
-            } else {
-                Ok(Some(Value::Null))
-            }
-        } else {
-            Ok(Some(Value::Null))
-        }
-    }
+    // Note: handle_document_link is implemented in language/misc.rs
 }
 
 impl LspServer {
